Turn Methods into a Comma Separated List
The pairing methods is a list of lists right now, which makes it annoying to parse.
I don't see a strict reason for the separation of lists, as we are just making sure that each method is registered. The URI doesn't necessarily know what the difference between them is either.

So I suggest changing it in this way to make it easier to parse, and construct.

I'm open to being refuted here.